+++ GUI-dev/src/app/@theme/providers/backend-server/gateway.ts
@@ -58,7 +58,34 @@ export class GatewayProvider {
   }
 
   postStoreProcessRatings(processId: string, rating: StoreProcessRating): void {
-    const url = this.serverConfig.postStoreProcessRating + '/' + processId
+    const url = this.serverConfig.postStoreProcessRating + '/' + processId;
     this.http.post<StoreProcessRating>(url, rating).toPromise()
   }
+
+  getUnapprovedStoreProcesses(): Promise<StoreProcess[]> {
+    return this.http.get<StoreProcess[]>(this.serverConfig.getUnapprovedStoreProcesses)
+      .toPromise()
+  }
+
+  postStoreProcessApproved(processId: string): void {
+    const url = this.serverConfig.postStoreProcessApproved + '/' + processId + '/approve';
+    this.http.post<StoreProcess>(url, processId).toPromise()
+  }
+
+  postStoreProcessUnapproved(processId: string): void {
+    const url = this.serverConfig.postStoreProcessUnapproved + '/' + processId + '/unapprove';
+    this.http.post<StoreProcess>(url, processId).toPromise()
+  }
+
+  getStoreProcessById(processId: string): Promise<StoreProcess> {
+    const url = this.serverConfig.getStoreProcessById + '/' + processId;
+    return this.http.get<StoreProcess>(url)
+      .toPromise()
+  }
+
+  postStoreProcessComment(comment: string, processId: string): void {
+    const url = this.serverConfig.postStoreProcessApproved + '/' + processId + '/updateApprovalComment';
+    this.http.post<StoreProcess>(url, comment).toPromise()
+  }
+
 }