+++ GUI-dev/src/app/pages/approval-details/approval-details.module.ts
@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
 
-
 import { ThemeModule } from '../../@theme/theme.module';
 
 
@@ -12,4 +11,4 @@ import { ThemeModule } from '../../@theme/theme.module';
     // ApprovalDetailsComponent,
   ],
 })
-export class ApprovalDetailsModule { }
+export class ApprovalDetailsModule {}